feat(graduate): 졸업 방식 신청 페이지 API 연동
