+++ apps/graduate/src/pages/client/apply/api/submitGraduationType.ts
@@ -1,0 +1,17 @@
+import { useMutation } from '@tanstack/react-query';
+import { patch } from '~/shared/api';
+import { END_POINT, type GraduationType } from '~/shared/constants';
+
+const submitGraduationType = async (graduationType: GraduationType) => {
+  const response = await patch({
+    request: END_POINT.USER.GRADUATION_TYPE,
+    data: { graduationType },
+  });
+  return response.data;
+};
+
+export const useSubmitGraduationType = () => {
+  return useMutation({
+    mutationFn: submitGraduationType,
+  });
+};
